Which versions of your npm package the world is using?
Added a tool to count which versions of a package are referenced
in the npm graph. Fixes https://github.com/anvaka/npmrank/issues/6

For now this is just a command line utility. It take some, since it
re-reads downloaded `byField` response on every iteration.

Ideally this tool will require a special graph where each node is a
package name and each edge is a dependency, with version information.
After this is done queries will run extremely fast.